backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
jinja2==3.1.4
python-dotenv==1.0.1
PyGithub==2.6.1
pydantic==2.11.5
aiofiles==23.2.1
anthropic>=0.52.2
pytest==8.1.1
httpx==0.27.0  # Required for TestClient 